pyproject.toml
[project]
name = "orion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "imagecodecs>=2026.1.14",
    "matplotlib>=3.10.8",
    "numpy>=2.2.2",
    "polars>=1.38.1",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.2",
    "scikit-image>=0.26.0",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "tifffile>=2025.1.10",
    "zarr>=3.1.5",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "ruff>=0.15.1",
]
